Add basic provision, deprovision, and update
We want the broker to be able to create and delete S3 buckets. We
provide the minimum required config at this stage: the region and an
optional prefix for the bucket name.

It is not clear at this stage what level of control over the bucket we
want to give to users. Therefore, update operations are explicitly not
supported.