Implement bind operation
The simplest way to implement authentication for a bucket without giving
the broker the ability to create IAM policies is to name a specific user
in a bucket policy. Therefore, we implement binding by creating an IAM
user and keys and putting a bucket policy which names them as the
`Principal`.

We have to put a retry mechanism into the putting of the policy because
IAM and S3 take a number of seconds to get in-sync with each other.

Also, we have to do some awkward type juggling because the policies are
returned via the Golang SDK as `*string` types, and the AWS Principal
may be a string or an array of strings, depending on whether there are
one or more existing principals.

Currently the S3 client integration tests leave an orphaned IAM user.
This should be fixed in an upcoming commit.