Allow bindings to be read-only or read-write
In this commit, we allow users to pass a "permissions" parameter when binding
an app or creating a service key, so that access credentials can be created
with as limited scope as necessary. The valid permissions are "read-only" and
"read-write".

We refactored how S3 bucket policies are generated and mutated, so that future
bucket policy changes are easier to implement.