Buckets can allow Public GetObject
A user can specify, at provision-time, whether a bucket should be
'public' (world-readable). The parameter for this is

    {"public_bucket": bool}

Any object can be accessed via a HTTP GET to the object's URL, but
listing of objects is not allowed.

This is implemented via a policy statement on the Bucket policy, in the
same way as a user created from a binding would be.