`costs_by_month` utility for retrieving S3 costs
This commit adds a utility, `costs_by_month`, that outputs the cost of
tenant S3 buckets over the last few months. Some of the costs are an
estimate rather than necessarily what we will be billed, but it gives
a good sense of whether our billing structure matches the costs to us.