Get and release locks when modifying a bucket
The S3 broker has race conditions wherein two competing requests can pull the
bucket policy from S3, modify it in different ways, and then push it back to
S3. This always results in one request or the other failing, and the backing
service getting in to a state which tenants cannot rectify for themselves.

This commit has the broker obtain a lock from Locket before performing any
operation on a bucket. This should eliminate the race condition between
requests, and the max 30s TTL on the lock should prevent it from slowing the S3
broker too much when multiple requests come in.