Unique owner for each lock for each request
Locket has the concept of an owner, which represents a thing which has
ownership of a lock

Previously this was the deployment, which is not unique between calls to
the s3 client operations. Therefore the locking didn't actually do
anything, because as far as Locket was concerned it was the same owner
renewing the lock, not two competing operations.

Instead we should generate a uuid for each operation and use that as the
owner, and log it accordingly

Co-authored-by: Andy Hunt <andy.hunt@digital.cabinet-office.gov.uk>
Co-authored-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>