Rework locket locking to work
Our understanding of the Locket API is once again wrong.

We expected it to hang for the TTL of the lock, which it does not do.

Instead it returns an error, we make a few attempts to grab the lock
(15) with 1s between attempts and the lock ttl is now only 10 seconds.

This means normal behaviour _should_ never hit a lock.

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>