Add integration test for distributed locking
We have an acceptance test in the pipeline but the feedback loop is
quite low.

The acceptance test uses real locket which is why it is good
but I've modified the locket stub here to kind of work for the same use
case. I hope to port this to the service-broker-base.

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>